milestone3/java/generated_code/task2/code26/src/main/java/org/example/App.java
package org.example;

import java.util.Arrays;
import java.util.List;


public class App {

    public static void main(String[] args) {
        // Create some example parameters for the createView method
        DataType dataType = new DataType();  // You would need to replace this with actual code to create a DataType
        Class<?> elementType = String.class;  // This is an example, replace with the actual class you want
        boolean isVlen = true;  // This is an example, replace with the actual value you want
        Index index = new Index();  // You would need to replace this with actual code to create an Index
        Object storage = new Object[] {"example1", "example2"};  // This is an example, replace with the actual storage you want

        // Call the createView method
        createView(dataType, elementType, isVlen, index, storage);

        // Do something with the result, for example print it
//        System.out.println(Arrays.toString(result));
    }

    protected static Object[] createView(DataType dataType, Class<?> elementType, boolean isVlen, Index index, Object storage) {
        return new List[]{ArrayObject.factory(dataType, elementType, isVlen, index, storage)};
    }
}

class DataType {
    private int elementSize;

    public DataType() {
        this.elementSize = elementSize;
    }

    public int getElementSize() {
        return elementSize;
    }
}

class Index {
    private int[] arr;

    public Index() {
        this.arr = arr;
    }

    public int[] getArr() {
        return arr;
    }
}

class ArrayObject {

    private DataType dataType;
    private Class<?> elementType;
    private boolean isVlen;
    private Index index;
    private Object storage;

    // Private constructor to prevent direct instantiation
    private ArrayObject(DataType dataType, Class<?> elementType, boolean isVlen, Index index, Object storage) {
        this.dataType = dataType;
        this.elementType = elementType;
        this.isVlen = isVlen;
        this.index = index;
        this.storage = storage;
    }

    // Factory method
    public static List factory(DataType dataType, Class<?> elementType, boolean isVlen, Index index, Object storage) {
        // Create an ArrayObject
        ArrayObject arrayObject = new ArrayObject(dataType, elementType, isVlen, index, storage);

        List<Object> list = null;
        if (storage instanceof Object[]) {
            list = Arrays.asList((Object[]) storage);
        }
        // Convert storage to an array, assuming it's a type that can be converted
        // You may need to adjust this depending on the actual type and contents of storage
        return list;
    }

    // Additional methods, getters, and setters would go here
}

