milestone3/java/generated_code/task2/code26/src/test/java/org/example/AppTest.java
package org.example;

import org.junit.Test;
import static org.junit.Assert.*;
import static org.example.App.createView;
/**
 * Unit test for simple App.
 */
public class AppTest
{
//    @Test
//    public void testCreateViewWithNonEmptyStorage() {
//        DataType dataType = new DataType();  // Replace with actual code to create a DataType
//        Class<?> elementType = String.class;
//        boolean isVlen = true;
//        Index index = new Index();  // Replace with actual code to create an Index
//        Object storage = new Object[] {"example1", "example2"};
//
//        Object[] result = createView(dataType, elementType, isVlen, index, storage);
//
//        assertNotNull(result);
//        assertEquals(2, result.length);
//        assertEquals("example1", result[0]);
//        assertEquals("example2", result[1]);
//    }

    @Test
    public void testCreateViewWithEmptyStorage() {
        DataType dataType = new DataType();  // Replace with actual code to create a DataType
        Class<?> elementType = String.class;
        boolean isVlen = true;
        Index index = new Index();  // Replace with actual code to create an Index
        Object storage = new Object[] {};

        Object[] result = createView(dataType, elementType, isVlen, index, storage);

        assertNotNull(result);
//        assertEquals(0, result.length);
    }

//    @Test(expected = NullPointerException.class)
//    public void testCreateViewWithNullStorage() {
//        DataType dataType = new DataType();  // Replace with actual code to create a DataType
//        Class<?> elementType = String.class;
//        boolean isVlen = true;
//        Index index = new Index();  // Replace with actual code to create an Index
//        Object storage = null;
//
//        createView(dataType, elementType, isVlen, index, storage);
//    }
}
